feat: like, dislike and hashtags highligth and redirect on posts
